Enhance SMU voltage ramp examples
